Skip `setup-custom2` in integration-tests2 for GHC >= 8.2
The `setup-custom2` tests implicit `setup-depends` which implies
a `Cabal < 2` upper bound, which is incompatible with GHC 8.2 or later

(cherry picked from commit 8947db4976a3bd38b0d652537c6efdacc3b6ea4a)

+++ cabal-install/tests/IntegrationTests2.hs
@@ -1347,14 +1347,16 @@ testSetupScriptStyles config reportSubCase = do
     marker1 @?= "ok"
     removeFile (basedir </> testdir1 </> "marker")
 
-    reportSubCase (show SetupCustomImplicitDeps)
-    (plan2, res2) <- executePlan =<< planProject testdir2 config
-    (pkg2,  _)    <- expectPackageInstalled plan2 res2 pkgidA
-    elabSetupScriptStyle pkg2 @?= SetupCustomImplicitDeps
-    hasDefaultSetupDeps pkg2 @?= Just True
-    marker2 <- readFile (basedir </> testdir2 </> "marker")
-    marker2 @?= "ok"
-    removeFile (basedir </> testdir2 </> "marker")
+    -- implicit deps implies 'Cabal < 2' which conflicts w/ GHC 8.2 or later
+    when (compilerVersion (pkgConfigCompiler sharedConfig) < mkVersion [8,2]) $ do
+      reportSubCase (show SetupCustomImplicitDeps)
+      (plan2, res2) <- executePlan =<< planProject testdir2 config
+      (pkg2,  _)    <- expectPackageInstalled plan2 res2 pkgidA
+      elabSetupScriptStyle pkg2 @?= SetupCustomImplicitDeps
+      hasDefaultSetupDeps pkg2 @?= Just True
+      marker2 <- readFile (basedir </> testdir2 </> "marker")
+      marker2 @?= "ok"
+      removeFile (basedir </> testdir2 </> "marker")
 
     reportSubCase (show SetupNonCustomInternalLib)
     (plan3, res3) <- executePlan =<< planProject testdir3 config
